[GORDO-1642] Try and address some obvious timewasters (#18970)
* Try and address some obvious timewasters

* More usual suspects

* Message batch size scaling

* move n reference

* Batch sizes are size_t.

* More size_t

+++ arangod/Pregel/Worker/State.h
@@ -98,7 +98,7 @@ struct WorkerState {
   std::unique_ptr<MessageCombiner<M>> messageCombiner;
   std::unique_ptr<InCache<M>> readCache = nullptr;
   std::unique_ptr<InCache<M>> writeCache = nullptr;
-  uint32_t messageBatchSize = 500;
+  size_t messageBatchSize = 500;
   std::unordered_map<ShardID, actor::ActorPID> responsibleActorPerShard;
 
   const actor::ActorPID conductor;
